Use "::" instead of ":" for pseudo-elements
CSS2 standards introduced pseudo-elements with the `:` syntax.

CSS3 standards revised the pseudo-element syntax.
According to CSS3 standards, pseudo-elements
should now use `::` instead of `:`.

I believe this book should use the `::` syntax for pseudo-elements.
This is in the interest of representing modern best-practices.

A counter-argument involves the wider browser-support of `:`.
While CSS3 introduced the new `::` syntax, it also required all user-agents
to maintain backwards-compatibility with the `:` syntax.
Therefore, `:` has wider-browser support.

One could argue that we should represent the syntax with wider
browser-support.

I think this counter-argument is weak because

1. `::` syntax is supported in IE 9+, Firefox 3.5+, Opera 12.1, Safari 5.1,
    and Chrome 9+
2. IE 8 has ~0.27% browser market-share. IE 6 and IE 7 have ~0.01%.
    Prior versions of FF, Opera, Safari, and Chrome have ~0%, ~0.01%,
    ~0.02%, and ~0.01% browser market-shares respectively.
3. Thus, the unsupportive browsers' cumulative market-share is
    negligible.
4. Using `::` syntax is important for distinguishing pseudo-elements
    from pseudo-classes, which still use the `:` syntax as of CSS3.
5. The same backwards-compatibility argument can be made for
    using floats instead of flexbox or grid. Yet, the consensus is that
    technological progress is more important than backwards-compatibility.
    I believe the same principle applies here.

(Source for browser market-share data: caniuse.com)